Fixed time axis issue
If the number of partitions in time axis weren't equal to number of years then
showing only year was confusing since 2010 could represent Jan 2010 or Nov 2010.
It was confusing when a ppoint with 2011 showed up immediately after 2010.
Now we show year axis only if we exactly as many years as time axis partitions.