Fixed tooltip placement on chart
The target value computation corresponding to tooltip position relied on data being evenly distributed. Fixed the logic to compute target value based on domain of axis.
The binary search logic that finds the finds the rows that match the value corresponding to tooltips position had an issue where it was always returning the lragest value less than target value.